adjusting get_pytest_results to separate from tests
